Added RGBA<T> class in a very ugly manner. Replaced RandomInt and RandomFloat by RandomRange. Added () and [] operators to CImageData.

+++ Engine/inc/2de_GraphicsHigh.h
@@ -74,7 +74,7 @@ public:
 	int					Emission;					//	Number of particles per frame or second
 
 	float				ParticleLife;
-	int					ParticleLifeVariability;
+	float				ParticleLifeVariability;
 
 	RGBAf				ColorStart;
 	RGBAf				ColorOver;
